celery_ipc/celeryconfig.py
import os

from dotenv import load_dotenv

load_dotenv()

RABBITMQ_USER = os.getenv("RABBITMQ_USER")
RABBITMQ_PASSWORD = os.getenv("RABBITMQ_PASSWORD")
RABBITMQ_HOST = os.getenv("RABBITMQ_HOST")
RABBITMQ_PORT = os.getenv("RABBITMQ_PORT")
CONNECTION_URI = (
    f"pyamqp://{RABBITMQ_USER}:{RABBITMQ_PASSWORD}@{RABBITMQ_HOST}:{RABBITMQ_PORT}/"
)

imports = ("tasks",)
accept_content = ["msgpack"]
broker_url = CONNECTION_URI
task_serializer = "msgpack"
result_serializer = "msgpack"
