pyproject.toml
[tool.poetry]
name = "kumiko-ipc"
version = "0.1.0"
description = "The IPC server for Kumiko"
authors = ["No767 <73260931+No767@users.noreply.github.com>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
python-dotenv = "^0.21.0"
better-ipc = "^2.0.1"
celery = {extras = ["gevent", "msgpack", "redis"], version = "^5.2.7"}
librabbitmq = {git = "https://github.com/celery/librabbitmq.git"}
amqp = "^5.1.1"
asgiref = "^3.5.2"
discord-py = "^2.1.0"
dramatiq = {extras = ["rabbitmq", "watch"], version = "^1.13.0"}

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"

[tool.isort]
profile = 'black'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
